refactor: use API RunAs and ConnectionState, remove internal duplicates
