+++ backends/folia/src/main/java/dev/objz/commandbridge/folia/FoliaExecutor.java
@@ -3,7 +3,7 @@ package dev.objz.commandbridge.folia;
 import dev.objz.commandbridge.backends.platform.PlatformExecutor;
 import dev.objz.commandbridge.logging.Log;
 import dev.objz.commandbridge.net.payloads.cmd.ExecuteCommand;
-import dev.objz.commandbridge.scripting.model.enums.RunAs;
+import dev.objz.commandbridge.api.channel.command.RunAs;
 import io.papermc.paper.threadedregions.scheduler.GlobalRegionScheduler;
 import io.papermc.paper.threadedregions.scheduler.EntityScheduler;
 import org.bukkit.Bukkit;
