+++ backends/velocity/src/main/java/dev/objz/commandbridge/velocity/VelocityExecutor.java
@@ -7,7 +7,7 @@ import com.velocitypowered.api.proxy.ProxyServer;
 import dev.objz.commandbridge.backends.platform.cmd.CommandExecutor;
 import dev.objz.commandbridge.logging.Log;
 import dev.objz.commandbridge.net.payloads.cmd.ExecuteCommand;
-import dev.objz.commandbridge.scripting.model.enums.RunAs;
+import dev.objz.commandbridge.api.channel.command.RunAs;
 import net.kyori.adventure.text.Component;
 
 import java.util.Objects;
