+++ core/build.gradle.kts
@@ -15,6 +15,7 @@ repositories {
 }
 
 dependencies {
+    api(project(":api"))
     api("com.fasterxml.jackson.core:jackson-databind:2.18.2")
     api("com.fasterxml.jackson.core:jackson-annotations:2.18")
     api("com.fasterxml.jackson.core:jackson-core:2.18.2")
