+++ core/src/main/java/dev/objz/commandbridge/net/payloads/cmd/ExecuteCommand.java
@@ -2,7 +2,7 @@ package dev.objz.commandbridge.net.payloads.cmd;
 
 import java.util.Set;
 import java.util.UUID;
-import dev.objz.commandbridge.scripting.model.enums.RunAs;
+import dev.objz.commandbridge.api.channel.command.RunAs;
 
 public record ExecuteCommand(
         String command,
