--- core/src/main/java/dev/objz/commandbridge/scripting/model/enums/RunAs.java
@@ -1,7 +1,0 @@
-package dev.objz.commandbridge.scripting.model.enums;
-
-public enum RunAs {
-    CONSOLE,
-    PLAYER,
-    OPERATOR
-}
